# Split compiled code and export data for each archive