This a PoC to showcase #1803. Tests are not passing, but builds are. I'll update the tests asap.

It's pretty nice.

The metadata archive and the linking archives are smaller than the full archive, this means:
- most actions are faster
- linking is faster since the linking archive is smaller than the full archive
- i would figure this leads to significant speedups in RBE

All in all, in my non trivial project, building an Android AAR, changing a deep package:

Don't look at the time too much, as least half is spent zipping the AAR.

Without this PR:
```
INFO: Elapsed time: 15.596s, Critical Path: 14.72s
INFO: 22 processes: 22 darwin-sandbox.
```

With this PR:
```
INFO: Elapsed time: 12.385s, Critical Path: 11.73s
INFO: 6 processes: 6 darwin-sandbox.
```