@steeve 2019-08-23
More benchmarks, from a widely used package in our repo:
```
INFO: Elapsed time: 17.364s, Critical Path: 14.68s
INFO: 57 processes: 57 darwin-sandbox.
```

```
INFO: Elapsed time: 13.916s, Critical Path: 13.56s
INFO: 6 processes: 6 darwin-sandbox.
```