@steeve 2019-08-23
Here is a quick view of the sizes for huge gRPC proto file we have internally:
```
-r-xr-xr-x  1 steeve  wheel   501K Aug 24 01:28 obfs.a
-r-xr-xr-x  1 steeve  wheel    14M Aug 24 01:28 obfs.link.a
```