@steeve 2019-08-24
Everything seems to be working, save for plugin mode. Apparently it needs the full link file.

I'm looking into disabling linkobj in `plugin` mode.