@steeve 2019-08-24
One thing we can do is, if the mode is `plugin`, at link, merge the `.a` and `.link.a` and pass that to the linker. That way the benefit is kept.