@steeve 2019-08-24
Ok so if a binary imports `plugin`, then it can't link with `.link.a` archives, it needs the full archives. The problem is that it can't be detected in `go.mode`, hence why the test is failing