@steeve 2019-08-24
Since the plugin use case is so niche, I'm wondering if having an attribute for binaries that use `import "plugin"` would be useful, and it would disable linkobj mode (which the vast majority of users would benefit from)