@steeve 2019-08-24
Thinking out loud, @jayconrod, since you have access to the google rules, how do they deal with that case?