@steeve 2019-09-02
So after discussing on https://github.com/golang/go/issues/33820, here are the options:
- make the feature opt-in (perhaps via `features`), and thus make it part of the `go.mode`
- make the feature opt-out, since I would figure not many people use `plugin` anyway