@steeve 2019-09-05
I've been thinking perhaps the best way is to do what rules_apple does at https://github.com/bazelbuild/rules_apple/blob/master/doc/common_info.md

Something like:
```
--define=io.bazel.rules_go.linkobj=1
--define=go.linkobj=1
```