@steeve 2019-09-05
So I went ahead and added opt-in via `--define=go.linkobj=1`. I'll update the README if you guys are fine with the solution.